gmsol_timelock

Module cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

Return

Functions

approve_instruction

approve_instructions

cancel_instruction

cancel_instructions

create_instruction_buffer

execute_instruction

increase_delay

initialize_config

initialize_executor

revoke_role

set_expected_price_provider