gmsol_timelock::cpi

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

ApproveInstruction

Generated CPI struct of the accounts for ApproveInstruction.

ApproveInstructions

Generated CPI struct of the accounts for ApproveInstructions.

CancelInstruction

Generated CPI struct of the accounts for CancelInstruction.

CancelInstructions

Generated CPI struct of the accounts for CancelInstructions.

CreateInstructionBuffer

Generated CPI struct of the accounts for CreateInstructionBuffer.

ExecuteInstruction

Generated CPI struct of the accounts for ExecuteInstruction.

IncreaseDelay

Generated CPI struct of the accounts for IncreaseDelay.

InitializeConfig

Generated CPI struct of the accounts for InitializeConfig.

InitializeExecutor

Generated CPI struct of the accounts for InitializeExecutor.

RevokeRole

Generated CPI struct of the accounts for RevokeRole.

SetExpectedPriceProvider

Generated CPI struct of the accounts for SetExpectedPriceProvider.